move test migrate
